Update README for v4.2 terp
I've now ported version 4.2 of the interpreter ("terp") to BBC BASIC. Version 4.2 of the interpreter is the version that was originally published in SoftSide magazine in 1980. v4.2 of the terp correctly interprets and plays the Adventureland game-data as it was published in SoftSide, which isn't surprising given that the game-data in SoftSide was expressly designed to be used with a v4.2 terp!